+++ tests/TestCase/View/Helper/MenuHelperTest.php
@@ -38,9 +38,9 @@ class MenuHelperTest extends HelperTestCase
      */
     protected function buildLinks($links)
     {
-        return array_map(function ($link) {
+        return implode(PHP_EOL, array_map(function ($link) {
             return call_user_func_array([$this->getMockForHelper(HtmlHelper::class, null), 'link'], $link);
-        }, $links);
+        }, $links));
     }
 
     /**
@@ -49,25 +49,19 @@ class MenuHelperTest extends HelperTestCase
      */
     public function testPosts()
     {
-        list($links, $title, $options) = $this->Helper->posts();
-        $this->assertEquals([
-            '<a href="/admin" title="List posts">List posts</a>',
-            '<a href="/me-cms/admin/posts/add" title="Add post">Add post</a>',
-            '<a href="/me-cms/admin/posts-tags" title="List tags">List tags</a>',
-        ], $this->buildLinks($links));
-        $this->assertEquals(I18N_POSTS, $title);
-        $this->assertEquals(['icon' => 'far file-alt'], $options);
+        list($links,,, $handledControllers) = $this->Helper->posts();
+        $links = $this->buildLinks($links);
+        $this->assertNotEmpty($links);
+        $this->assertTextNotContains('List categories', $links);
+        $this->assertTextNotContains('Add category', $links);
+        $this->assertEquals(['Posts', 'PostsCategories', 'PostsTags'], $handledControllers);
 
-        $expected = [
-            '<a href="/admin" title="List posts">List posts</a>',
-            '<a href="/me-cms/admin/posts/add" title="Add post">Add post</a>',
-            '<a href="/me-cms/admin/posts-categories" title="List categories">List categories</a>',
-            '<a href="/me-cms/admin/posts-categories/add" title="Add category">Add category</a>',
-            '<a href="/me-cms/admin/posts-tags" title="List tags">List tags</a>',
-        ];
         foreach (['manager', 'admin'] as $name) {
             $this->writeAuthOnSession(['group' => compact('name')]);
-            $this->assertEquals($expected, $this->buildLinks($this->Helper->posts()[0]));
+            list($links) = $this->Helper->posts();
+            $links = $this->buildLinks($links);
+            $this->assertTextContains('List categories', $links);
+            $this->assertTextContains('Add category', $links);
         }
     }
 
@@ -77,24 +71,19 @@ class MenuHelperTest extends HelperTestCase
      */
     public function testPages()
     {
-        list($links, $title, $options) = $this->Helper->pages();
-        $this->assertEquals([
-            '<a href="/me-cms/admin/pages" title="List pages">List pages</a>',
-            '<a href="/me-cms/admin/pages/index-statics" title="List static pages">List static pages</a>',
-        ], $this->buildLinks($links));
-        $this->assertEquals(I18N_PAGES, $title);
-        $this->assertEquals(['icon' => 'far copy'], $options);
+        list($links,,, $handledControllers) = $this->Helper->pages();
+        $links = $this->buildLinks($links);
+        $this->assertNotEmpty($links);
+        $this->assertTextNotContains('List categories', $links);
+        $this->assertTextNotContains('Add category', $links);
+        $this->assertEquals(['Pages', 'PagesCategories'], $handledControllers);
 
-        $expected = [
-            '<a href="/me-cms/admin/pages" title="List pages">List pages</a>',
-            '<a href="/me-cms/admin/pages/add" title="Add page">Add page</a>',
-            '<a href="/me-cms/admin/pages-categories" title="List categories">List categories</a>',
-            '<a href="/me-cms/admin/pages-categories/add" title="Add category">Add category</a>',
-            '<a href="/me-cms/admin/pages/index-statics" title="List static pages">List static pages</a>',
-        ];
         foreach (['manager', 'admin'] as $name) {
             $this->writeAuthOnSession(['group' => compact('name')]);
-            $this->assertEquals($expected, $this->buildLinks($this->Helper->pages()[0]));
+            list($links) = $this->Helper->pages();
+            $links = $this->buildLinks($links);
+            $this->assertTextContains('List categories', $links);
+            $this->assertTextContains('Add category', $links);
         }
     }
 
@@ -104,22 +93,9 @@ class MenuHelperTest extends HelperTestCase
      */
     public function testPhotos()
     {
-        list($links, $title, $options) = $this->Helper->photos();
-
-        $expected = [
-            '<a href="/me-cms/admin/photos" title="List photos">List photos</a>',
-            '<a href="/me-cms/admin/photos/upload" title="Upload photos">Upload photos</a>',
-            '<a href="/me-cms/admin/photos-albums" title="List albums">List albums</a>',
-            '<a href="/me-cms/admin/photos-albums/add" title="Add album">Add album</a>',
-        ];
-        $this->assertEquals($expected, $this->buildLinks($links));
-        $this->assertEquals(I18N_PHOTOS, $title);
-        $this->assertEquals(['icon' => 'camera-retro'], $options);
-
-        foreach (['manager', 'admin'] as $name) {
-            $this->writeAuthOnSession(['group' => compact('name')]);
-            $this->assertEquals($expected, $this->buildLinks($this->Helper->photos()[0]));
-        }
+        list($links,,, $handledControllers) = $this->Helper->photos();
+        $this->assertNotEmpty($this->buildLinks($links));
+        $this->assertEquals(['Photos', 'PhotosAlbums'], $handledControllers);
     }
 
     /**
@@ -128,24 +104,21 @@ class MenuHelperTest extends HelperTestCase
      */
     public function testBanners()
     {
-        $this->assertNull($this->Helper->banners());
-
-        $expected = [
-            '<a href="/me-cms/admin/banners" title="List banners">List banners</a>',
-            '<a href="/me-cms/admin/banners/upload" title="Upload banners">Upload banners</a>',
-        ];
+        $this->assertEmpty($this->Helper->banners());
 
         $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
-        list($links, $title, $options) = $this->Helper->banners();
-        $this->assertEquals($expected, $this->buildLinks($links));
-        $this->assertEquals('Banners', $title);
-        $this->assertEquals(['icon' => 'shopping-cart'], $options);
+        list($links,,, $handledControllers) = $this->Helper->banners();
+        $links = $this->buildLinks($links);
+        $this->assertNotEmpty($links);
+        $this->assertTextNotContains('List positions', $links);
+        $this->assertTextNotContains('Add position', $links);
+        $this->assertEquals(['Banners', 'BannersPositions'], $handledControllers);
 
-        $expected[] = '<a href="/me-cms/admin/banners-positions" title="List positions">List positions</a>';
-        $expected[] = '<a href="/me-cms/admin/banners-positions/add" title="Add position">Add position</a>';
         $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
         list($links) = $this->Helper->banners();
-        $this->assertEquals($expected, $this->buildLinks($links));
+        $links = $this->buildLinks($links);
+        $this->assertTextContains('List positions', $links);
+        $this->assertTextContains('Add position', $links);
     }
 
     /**
@@ -154,24 +127,21 @@ class MenuHelperTest extends HelperTestCase
      */
     public function testUsers()
     {
-        $this->assertNull($this->Helper->users());
-
-        $expected = [
-            '<a href="/me-cms/admin/users" title="List users">List users</a>',
-            '<a href="/me-cms/admin/users/add" title="Add user">Add user</a>',
-        ];
+        $this->assertEmpty($this->Helper->users());
 
         $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
-        list($links, $title, $options) = $this->Helper->users();
-        $this->assertEquals($expected, $this->buildLinks($links));
-        $this->assertEquals('Users', $title);
-        $this->assertEquals(['icon' => 'users'], $options);
+        list($links,,, $handledControllers) = $this->Helper->users();
+        $links = $this->buildLinks($links);
+        $this->assertNotEmpty($links);
+        $this->assertTextNotContains('List groups', $links);
+        $this->assertTextNotContains('Add group', $links);
+        $this->assertEquals(['Users', 'UsersGroups'], $handledControllers);
 
-        $expected[] = '<a href="/me-cms/admin/users-groups" title="List groups">List groups</a>';
-        $expected[] = '<a href="/me-cms/admin/users-groups/add" title="Add group">Add group</a>';
         $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
         list($links) = $this->Helper->users();
-        $this->assertEquals($expected, $this->buildLinks($links));
+        $links = $this->buildLinks($links);
+        $this->assertTextContains('List groups', $links);
+        $this->assertTextContains('Add group', $links);
     }
 
     /**
@@ -180,19 +150,15 @@ class MenuHelperTest extends HelperTestCase
      */
     public function testBackups()
     {
-        $this->assertNull($this->Helper->backups());
+        $this->assertEmpty($this->Helper->backups());
 
         $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
         $this->assertNull($this->Helper->backups());
 
         $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
-        list($links, $title, $options) = $this->Helper->backups();
-        $this->assertEquals([
-            '<a href="/me-cms/admin/backups" title="List backups">List backups</a>',
-            '<a href="/me-cms/admin/backups/add" title="Add backup">Add backup</a>',
-        ], $this->buildLinks($links));
-        $this->assertEquals('Backups', $title);
-        $this->assertEquals(['icon' => 'database'], $options);
+        list($links,,, $handledControllers) = $this->Helper->backups();
+        $this->assertNotEmpty($this->buildLinks($links));
+        $this->assertEquals(['Backups'], $handledControllers);
     }
 
     /**
@@ -201,26 +167,16 @@ class MenuHelperTest extends HelperTestCase
      */
     public function testSystems()
     {
-        $this->assertNull($this->Helper->systems());
+        $this->assertEmpty($this->Helper->systems());
 
         $this->writeAuthOnSession(['group' => ['name' => 'manager']]);
-        list($links, $title, $options) = $this->Helper->systems();
-        $this->assertEquals([
-            '<a href="/me-cms/admin/systems/tmp-viewer" title="Temporary files">Temporary files</a>',
-            '<a href="/me-cms/admin/systems/checkup" title="System checkup">System checkup</a>',
-            '<a href="/me-cms/admin/systems/browser" title="Media browser">Media browser</a>',
-            '<a href="/me-cms/admin/systems/changelogs" title="Changelogs">Changelogs</a>',
-        ], $this->buildLinks($links));
-        $this->assertEquals('System', $title);
-        $this->assertEquals(['icon' => 'wrench'], $options);
+        list($links,,, $handledControllers) = $this->Helper->systems();
+        $this->assertNotEmpty($this->buildLinks($links));
+        $this->assertTextNotContains('Log management', $links);
+        $this->assertEquals(['Logs', 'Systems'], $handledControllers);
 
         $this->writeAuthOnSession(['group' => ['name' => 'admin']]);
-        $this->assertEquals([
-            '<a href="/me-cms/admin/systems/tmp-viewer" title="Temporary files">Temporary files</a>',
-            '<a href="/me-cms/admin/systems/checkup" title="System checkup">System checkup</a>',
-            '<a href="/me-cms/admin/systems/browser" title="Media browser">Media browser</a>',
-            '<a href="/me-cms/admin/systems/changelogs" title="Changelogs">Changelogs</a>',
-            '<a href="/me-cms/admin/logs" title="Log management">Log management</a>',
-        ], $this->buildLinks($this->Helper->systems()[0]));
+        list($links) = $this->Helper->systems();
+        $this->assertTextContains('Log management', $this->buildLinks($links));
     }
 }